requirements.txt
Flask
Flask-Cors
arcgis
gunicorn
python-dotenv
pandas
geopandas